+++ config/environments/production.rb
@@ -59,7 +59,8 @@ Rails.application.configure do
 
   # Set host to be used by links generated in mailer templates.
   config.action_mailer.default_url_options = { protocol: "https", host: "heimdall.test" }
-
+  config.action_mailer.delivery_method = :mailersend
+  
   # Specify outgoing SMTP server. Remember to add smtp/* credentials via bin/rails credentials:edit.
   # config.action_mailer.smtp_settings = {
   #   user_name: Rails.application.credentials.dig(:smtp, :user_name),